# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/events": {
            "get": {
                "description": "Retrieves a list of all events with pagination",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Event"
                ],
                "summary": "Get all events",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page number for pagination",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of items per page for pagination",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of events",
                        "schema": {
                            "$ref": "#/definitions/bikefest_pkg_model.EventListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/bikefest_pkg_model.Response"
                        }
                    }
                }
            }
        },
...
        "/events/{id}": {
            "get": {
                "description": "Retrieves an event by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Event"
                ],
                "summary": "Get an event by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/bikefest_pkg_model.EventResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/bikefest_pkg_model.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/bikefest_pkg_model.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Updates an event by ID with new details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Event"
                ],
                "summary": "Update an event",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Event Update Information",
                        "name": "event",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/bikefest_pkg_model.CreateEventRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Event successfully updated",
                        "schema": {
                            "$ref": "#/definitions/bikefest_pkg_model.EventResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request - Invalid input",
                        "schema": {
                            "$ref": "#/definitions/bikefest_pkg_model.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/bikefest_pkg_model.Response"
                        }
                    }
                }
            }
        },
...
        "/users/events": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieves a list of events associated with a user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get User Events",
                "responses": {
                    "200": {
                        "description": "List of events associated with the user",
                        "schema": {
                            "$ref": "#/definitions/bikefest_pkg_model.EventListResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/bikefest_pkg_model.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Subscribes a user to an event with the provided details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Subscribe to an event",
                "parameters": [
                    {
                        "description": "Event Subscription Request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/bikefest_pkg_model.CreateEventRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully subscribed to the event",
                        "schema": {
                            "$ref": "#/definitions/bikefest_pkg_model.EventResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request - Invalid input, such as invalid time format or missing required fields",
                        "schema": {
                            "$ref": "#/definitions/bikefest_pkg_model.Response"
                        }
                    },
                    "409": {
                        "description": "Conflict - User already subscribed to the event",
                        "schema": {
                            "$ref": "#/definitions/bikefest_pkg_model.Response"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity - User has exceeded the maximum number of subscriptions",
                        "schema": {
                            "$ref": "#/definitions/bikefest_pkg_model.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error - Error storing the event, subscribing the user, or enqueuing the event notification",
                        "schema": {
                            "$ref": "#/definitions/bikefest_pkg_model.Response"
                        }
                    }
                }
            }
        },
...
        "/users/events/{event_id}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Deletes a specific event by its ID for a given user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Delete event",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Event ID",
                        "name": "event_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Event successfully deleted",
                        "schema": {
                            "$ref": "#/definitions/bikefest_pkg_model.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/bikefest_pkg_model.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "bikefest_pkg_model.CreateEventRequest": {
            "type": "object",
            "properties": {
                "event_detail": {
                    "type": "string",
                    "example": "{\"title\":\"test event\",\"description\":\"test event description\"}"
                },
                "event_time_end": {
                    "type": "string",
                    "example": "2021/01/01 00:00"
                },
                "event_time_start": {
                    "type": "string",
                    "example": "2021/01/01 00:00"
                },
                "id": {
                    "type": "string"
                }
            }
        },
...
        "bikefest_pkg_model.Event": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "event_detail": {
                    "description": "the `EventDetail` field store the event detail in json format, this would be parsed when send to line message API",
                    "type": "string"
                },
                "event_time_end": {
                    "type": "string"
                },
                "event_time_start": {
                    "type": "string"
                },
                "id": {
                    "description": "the event id is defne at the frontend, if frontend don't have event id, the event id would be calculated by the hash of event detail and event time",
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "bikefest_pkg_model.EventListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/bikefest_pkg_model.Event"
                    }
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "bikefest_pkg_model.EventResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/bikefest_pkg_model.Event"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "bikefest_pkg_model.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
...
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}